Tech: extrait le code du formulaire de choix du test de dépistage
